Kunal2

See merge request compsci308_2019spring/voogasalad_nomergeconflicts!135

+++ src/Configurations/src/Configs/ArsenalConfig/WeaponUnlockerBehaviors/LevelUnlockable.java
@@ -16,7 +16,7 @@ public class LevelUnlockable extends ArsenalBehavior implements Updatable {
     }
 
     @Override
-    public void update(long ms) {
+    public void update(double ms) {
 
     }
